requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
postgresql==1.0.7
pydantic-settings==2.0.3
fastapi-mail==1.4.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
alembic==1.12.1
python-multipart==0.0.6
Jinja2==3.1.2
websockets==12.0
argon2-cffi==23.1.0
boto3==1.28.85
aiohttp==3.9.1
requests==2.31.0
slowapi==0.1.9
pytest==7.4.0
pytest-asyncio==0.21.0